requirements.txt
rumps>=0.4.0
pynput>=1.7.6
sounddevice>=0.4.6
numpy>=1.24.0
pywhispercpp>=1.2.0
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-Quartz>=10.0
Pillow>=10.0.0
